Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-url-02-expected.txt

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/plugins/resources/mock-plugin.pl?url-doesnt-match-csp because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/plugins/resources/mock-plugin-unknown-type.pl?type-attribute-doesnt-match-csp because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
This tests our handling of non-`data:` URLs, given a `plugin-types` CSP directive. Consider this test passing if none of the following frames contains "FAIL" and four sets of console logs appear above.



--------
Frame: '<!--frame1-->'
--------


--------
Frame: '<!--frame2-->'
--------


--------
Frame: '<!--frame3-->'
--------
